remove old BaseKafkaSpec, rename new one
